requirements.txt
fastapi
uvicorn
pymysql
tortoise-orm
pyjwt
aiomysql
pydantic
python-multipart
pillow
imagehash